chore: rename classifiers/intents/intent_model.py/KerasIntentModel to classifiers/keras_classification_model.py/KerasClassificationModel
