+++ deeppavlov/configs/intents/intents_dstc2.json
@@ -63,7 +63,7 @@
           "y_probas_dict"
         ],
         "main": true,
-        "name": "intent_model",
+        "name": "keras_classification_model",
         "save_path": "intents/intent_cnn_v4",
         "load_path": "intents/intent_cnn_v4",
         "classes": "#classes_vocab.keys()",
