+++ deeppavlov/configs/intents/intents_snips.json
@@ -60,7 +60,7 @@
           "y_probas_dict"
         ],
         "main": true,
-        "name": "intent_model",
+        "name": "keras_classification_model",
         "save_path": "intents/intent_cnn_snips_v4",
         "load_path": "intents/intent_cnn_snips_v4",
         "classes": "#classes_vocab.keys()",
