+++ deeppavlov/models/classifiers/keras_classification_model.py
@@ -29,9 +29,9 @@ from keras.regularizers import l2
 from deeppavlov.core.common.errors import ConfigError
 from deeppavlov.core.common.registry import register
 from deeppavlov.core.models.keras_model import KerasModel
-from deeppavlov.models.classifiers.intents.utils import labels2onehot, proba2labels
+from deeppavlov.models.classifiers.utils import labels2onehot, proba2labels
+from deeppavlov.models.classifiers.utils import md5_hashsum
 from deeppavlov.models.embedders.fasttext_embedder import FasttextEmbedder
-from deeppavlov.models.classifiers.intents.utils import md5_hashsum
 from deeppavlov.models.tokenizers.nltk_tokenizer import NLTKTokenizer
 from deeppavlov.core.common.log import get_logger
 from deeppavlov.core.layers.keras_layers import additive_self_attention, multiplicative_self_attention
@@ -40,8 +40,8 @@ from deeppavlov.core.layers.keras_layers import additive_self_attention, multipl
 log = get_logger(__name__)
 
 
-@register('intent_model')
-class KerasIntentModel(KerasModel):
+@register('keras_classification_model')
+class KerasClassificationModel(KerasModel):
     """
     Class implements keras model for intent recognition task for multi-class multi-label data
     """
