+++ requirements.txt
@@ -3,7 +3,7 @@ numpy==1.14.3
 lxml==4.1.1
 tqdm==4.19.5
 requests==2.18.4
-tensorflow==1.8.0
+tensorflow-gpu==1.8.0
 overrides==1.9
 kenlm==0.0.0
 h5py==2.7.1
